+++ .circleci/doTests.sh
@@ -65,7 +65,7 @@ then
 fi
 
 someTestsRan=false
-while read line
+while read -u 10 line
 do
     if [[ $line = "" ]]; then
         continue
@@ -155,7 +155,7 @@ do
         cd project/.circleci
         ./stopDb.sh $currPinnedDb
     done
-done <<< `cat pluginInterfaceExactVersionsOutput`
+done 10<pluginInterfaceExactVersionsOutput
 
 if [[ $someTestsRan = "true" ]]
 then