+++ .circleci/testCli.sh
@@ -12,6 +12,7 @@ sed -i 's/# refresh_api_path:/refresh_api_path: \/refresh/g' /usr/lib/supertoken
 sed -i 's/# cookie_domain:/cookie_domain: supertokens.io/g' /usr/lib/supertokens/config.yaml
 sed -i 's/# mysql_user:/mysql_user: root/g' /usr/lib/supertokens/config.yaml
 sed -i 's/# mysql_password:/mysql_password: root/g' /usr/lib/supertokens/config.yaml
+sed -i 's/# mongodb_connection_uri:/mongodb_connection_uri: mongodb:\/\/root:root@localhost:27017/g' /usr/lib/supertokens/config.yaml
 
 supertokens start dev --port=8888