+++ src/test/java/io/supertokens/test/SessionTest1.java
@@ -107,6 +107,10 @@ public class SessionTest1 {
         assertEquals(StorageLayer.getStorageLayer(process.getProcess()).getNumberOfSessions(), 1);
         assert sessionInfo.accessToken != null;
         assertNotNull(sessionInfo.antiCsrfToken);
+        assert sessionInfo.idRefreshToken != null;
+        assert sessionInfo.idRefreshToken.cookieSecure != null;
+        assert sessionInfo.idRefreshToken.cookiePath != null;
+        assert sessionInfo.idRefreshToken.domain != null;
 
         SessionInformationHolder verifiedSession = Session.getSession(process.getProcess(),
                 sessionInfo.accessToken.token, sessionInfo.antiCsrfToken, true);