+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2020-04-07
+### Added
+- Compatibility with NoSQL databases like MongoDB
+- Setting sameSite cookie option. However, this is not usable in this release.
+
 ## [1.1.1] - 2020-03-23
 ### Changed
 - Adds #!/bin/bash in scripts